+++ packages/language-js/src/expressions/__tests__/performance/query-selectors.test.ts
@@ -3,29 +3,28 @@ import type { MangleExpression } from "@webmangler/types";
 import { benchmarkFn, getRuntimeBudget } from "@webmangler/benchmarking";
 import { expect } from "chai";
 
-import { embedContentInContext } from "./benchmark-helpers";
+import { embedContentInContext } from "../common";
 
-import cssDeclarationPropertyExpressionFactory from "../css-properties";
+import querySelectorExpressionFactory from "../../query-selectors";
 
-suite("JavaScript - CSS Property Expression Factory", function() {
+suite("JavaScript - Query Selector Expression Factory", function() {
   let expressions: Iterable<MangleExpression>;
 
-  const patterns = "var-[a-zA-Z0-9-]+";
+  const patterns = "cls-[a-zA-Z0-9-]+";
 
-  const contentWithProperties = embedContentInContext(`
-    $element.style.getPropertyValue("--var-color");
-    $element.style.getPropertyValue("--var-font-size");
-    $element.style.getPropertyValue("--margin-left");
+  const contentWithQuerySelector = embedContentInContext(`
+    const div = document.querySelectorAll("div");
+    const foo = document.querySelectorAll(".cls-foo");
+    const bar = document.querySelectorAll("#bar");
   `);
-  const contentWithoutProperties = `
+  const contentWithoutQuerySelector = `
     const foo = "bar";
     const fooEl = document.getElementById(foo);
-    const barEl = document.querySelector(".bar");
   `;
 
   suiteSetup(function() {
-    expressions = cssDeclarationPropertyExpressionFactory({
-      prefix: "--",
+    expressions = querySelectorExpressionFactory({
+      prefix: "\\.",
     });
   });
 
@@ -35,7 +34,7 @@ suite("JavaScript - CSS Property Expression Factory", function() {
 
   test("simple file", function() {
     const budget = getRuntimeBudget(0.1);
-    const fileContent = contentWithProperties;
+    const fileContent = contentWithQuerySelector;
 
     let found: string[] = [];
     const result = benchmarkFn({
@@ -53,7 +52,7 @@ suite("JavaScript - CSS Property Expression Factory", function() {
 
   test("large file", function() {
     const budget = getRuntimeBudget(3);
-    const fileContent = contentWithProperties.repeat(100);
+    const fileContent = contentWithQuerySelector.repeat(100);
 
     let found: string[] = [];
     const result = benchmarkFn({
@@ -69,9 +68,9 @@ suite("JavaScript - CSS Property Expression Factory", function() {
     expect(result.medianDuration).to.be.below(budget);
   });
 
-  test("large file without properties", function() {
+  test("large file without query selectors", function() {
     const budget = getRuntimeBudget(1);
-    const fileContent = contentWithoutProperties.repeat(100);
+    const fileContent = contentWithoutQuerySelector.repeat(100);
 
     let found: string[] = [];
     const result = benchmarkFn({