+++ packages/language-js/CHANGELOG.md
@@ -8,6 +8,14 @@ Versioning].
 
 ## [Unreleased]
 
+- _No changes yet_
+
+## [0.1.26] - 2022-01-15
+
+- Fix query selectors not being mangled due to escaped quotes. ([#237])
+
+## [0.1.25] - 2022-01-08
+
 - Fix unintentional mangling in inline comments with newlines. ([#224])
 
 ## [0.1.24] - 2021-09-05
@@ -23,5 +31,6 @@ Versioning].
 [#136]: https://github.com/ericcornelissen/webmangler/pull/136
 [#139]: https://github.com/ericcornelissen/webmangler/pull/139
 [#224]: https://github.com/ericcornelissen/webmangler/pull/224
+[#237]: https://github.com/ericcornelissen/webmangler/pull/237
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/ "Keep a CHANGELOG"
 [semantic versioning]: https://semver.org/spec/v2.0.0.html "Semantic versioning"